Changes to all service makesfiles: Creation of xml documentation using go-unit-report and gocover-cobertura, which are displayed by when running tests of the individual services in azure pipelines. The installation of the required libraries takes place in the pipeline itself. Change to the order of tasks for integrationcheck and deployment (clean at the beginning instead of the end)
Change in the root top layer Makefile: The folders (services) in which actual changes were made in a feature branch are filtered using git diff. This speeds up the pipeline process, as an integration check is no longer carried out for all services.

Newly added: Pipeline file, it defines the triggers and individual steps that take place when either a pull request is performed or the main branch is merged. The changed service is also searched for here for the publication of the tests and code coverage using git diff.